requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
